+++ device-bootloaders/src/ServiceProcessorBL2/include/fru_defines.h
@@ -1,0 +1,22 @@
+#define FRU_BOARD_AREA 0
+#define FRU_PRODUCT    1
+#define BOARD_MFG      0
+#define BOARD_MFG      0
+#define BOARD_PNAME    1
+#define BOARD_SERIAL   2
+#define BOARD_PN       3
+#define BOARD_FILE     4
+#define PRODUCT_MFG    0
+#define PRODUCT_PNAME  1
+#define PRODUCT_PN     2
+#define PRODUCT_VER    3
+#define PRODUCT_SERIAL 4
+#define PRODUCT_ATAG   5
+#define PRODUCT_FILE   6
+
+#define BOARD_OFFSET_INDEX   3
+#define PRODUCT_OFFSET_INDEX 4
+
+#define FRU_INFO_MULTIPLIER 8
+#define BOARD_START_GAP     6
+#define PRODUCT_START_GAP   3